pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-daily-journal"
version = "0.1.0"
description = "AI Daily Journal: DB-first journaling with safe confirm workflow."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "argon2-cffi>=23.1.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pgvector>=0.3.5",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "sqlalchemy>=2.0.36",
  "typer>=0.12.5",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.2",
]

[project.scripts]
aijournal = "ai_daily_journal.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
